Options

This function creates a list of options for the use with the render function.

assemblerr_options(
  prm.use_mu_referencing = FALSE,
  ode.use_special_advans = TRUE,
  ode.use_general_linear_advans = TRUE,
  ode.general_nonlinear_advan = "advan13",
  ode.general_linear_advan = "advan5",
  ode.preferred_trans_routines = c("trans2", "trans4"),
  issues.missing_variables = c("fix-warn", "fix", "ignore", "fail")
)

Arguments

prm.use_mu_referencing

Use mu-referencing?

ode.use_special_advans

Use analytic solution ADVANs?

ode.use_general_linear_advans

Use ADVANs for linear ODEs?

ode.general_nonlinear_advan

ADVAN to be used for non-linear ODEs

ode.general_linear_advan

ADVAN to be used for linear ODEs

ode.preferred_trans_routines

Order of TRANS routines to be tried

issues.missing_variables

How to handle missing variables

Value

A list of options

Details

The function helps to create properly formatted list that can serve as input to the options= argument of the render() function.