PK distribution 2 compartments

This building block declares a two compartment distribution component for a pharmacokinetic model.

pk_distribution_2cmp(
  prm_vc = prm_log_normal("vc", median = 100, var_log = 0.1),
  prm_vp = prm_log_normal("vp", median = 5, var_log = 0.1),
  prm_q = prm_log_normal("q", median = 50, var_log = 0.1)
)

Arguments

prm_vc

Parameter model for the central volume of distribution

prm_vp

Parameter model for the peripheral volume of distribution

prm_q

Parameter model for the inter-compartmental clearance

Value

A building block of type 'pk_component'

Details

PK components

PK components can be added to a pk_model and exist in three different types: absorption, distribution, and elimination. The absorption component is optional, distribution and elimination are not and need to be added for the PK model to be valid.

A PK model can only have one component of each type and adding a component with an already existing type will replace the previous definition. For example, the distribution component will be a two compartment model in the following snippet:

pkm <- pk_model() +
  pk_absorption_fo() +
  pk_distribution_1cmp() +
  pk_distribution_2cmp() +
  pk_elimination_linear() +
  obs_additive(conc~C["central"]) 
pkm

Parameter models

All PK component functions allow the specification of the parameter model via their arguments. Arguments that refer to a parameter start with the prefix prm_. The default parameter model can be deduced from the default arguments in the usage section of the help entry. The parameter name, specified via the name= argument of the parameter model building block allows the renaming of the model parameters.

For example, the parameter prm_vc= refers to the central volume of distribution parameter in the one compartment distribution PK component and the default parameter model is a log-normal distribution. The following code block specifies a normal distribution parameter model and names the parameter v:

pk_distribution_1cmp(
    prm_vc = prm_normal("v", mean = 50, var = 25)
)

See also