Create a PK model

This function creates the basis for a pharmacokinetic model.

pk_model()

Value

A pk_model

Details

The function creates the foundation for a pharmacokinetic model to which different building blocks can be added. The following building blocks are relevant for this model type:

Parameters: prm_log_normal, prm_logit_normal, prm_no_var, prm_normal

Observations: obs_additive, obs_combined, obs_proportional

Algebraic relationships: algebraic

PK components: pk_absorption_fo, pk_absorption_fo_lag, pk_absorption_fo_transit, pk_absorption_fo_zo, pk_absorption_zo, pk_absorption_zo_lag, pk_distribution_1cmp, pk_distribution_2cmp, pk_distribution_3cmp, pk_elimination_linear, pk_elimination_linear_nl, pk_elimination_nl, pk_model

Input variables: input_variable, dataset