Model tasks

These functions define model tasks that can be combined using + and then be included during rendering.

tsk_estimation() adds a parameter estimation task allowing to specify the estimation algorithm to be used and whether a $COV step should be run.

tsk_ouput() adds an output task allowing to display model variables in a table.

tsk_ouput_xpose() adds several tables as output following the xpose4 output standard.

tsk_estimation(
  algorithm = c("foce", "foce-inter", "foce-no-inter", "fo", "imp", "saem"),
  se = FALSE,
  target_options = list()
)

tsk_output(filename = "sdtab", variables)

tsk_output_xpose4()

Arguments

algorithm
The estimation algorithm to use for the task

se
Whether to calculate parameter uncertainties

target_options
List of additional options that should be passed to NONMEM

filename
The filename for the output file

variables
The model variables that should be output

Examples

m <- model() +
  input_variable("dose") +
  prm_log_normal("emax", median = 10, var_log = 0.09) +
  prm_log_normal("ed50", median = 50, var_log = 0.09) +
  algebraic(effect~emax*dose/(ed50 + dose)) +
  obs_proportional(~effect, var_prop = 1)

# add estimation task using importance sampling, covariance step
# and user-defined ISAMPLE option
render(
  model = m,
  tasks = tsk_estimation(
    algorithm = "imp",
    se = TRUE,
    target_options = list(isample=1000)
  )
)

#> $PROBLEM
#> $INPUT DOSE ID DV
#> $DATA data.csv IGNORE=@
#> $PRED
#> EMAX = THETA(1) * EXP(ETA(1))
#> ED50 = THETA(2) * EXP(ETA(2))
#> EFFECT = EMAX * DOSE/(ED50 + DOSE)
#> Y = EFFECT + EFFECT * EPS(1)
#> $ESTIMATION METHOD=IMP AUTO=1 ISAMPLE=1000
#> $COVARIANCE PRINT=E
#> $THETA (0, 10, Inf) ; POP_EMAX
#> $THETA (0, 50, Inf) ; POP_ED50
#> $OMEGA 0.09 ; IIV_EMAX
#> $OMEGA 0.09 ; IIV_ED50
#> $SIGMA 1; RUV_PROP

# output model parameters to file 'prms'
render(m, tasks = tsk_output("prms", variables = vars_prms()))

#> $PROBLEM
#> $INPUT DOSE ID DV
#> $DATA data.csv IGNORE=@
#> $PRED
#> EMAX = THETA(1) * EXP(ETA(1))
#> ED50 = THETA(2) * EXP(ETA(2))
#> EFFECT = EMAX * DOSE/(ED50 + DOSE)
#> Y = EFFECT + EFFECT * EPS(1)
#> $TABLE EMAX ED50 FILE=prms NOAPPEND NOPRINT
#> $THETA (0, 10, Inf) ; POP_EMAX
#> $THETA (0, 50, Inf) ; POP_ED50
#> $OMEGA 0.09 ; IIV_EMAX
#> $OMEGA 0.09 ; IIV_ED50
#> $SIGMA 1; RUV_PROP